requirements.txt
# Core
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
tqdm==4.66.4
ujson==5.9.0
rich==13.7.1
pydantic==2.8.2
python-dotenv==1.0.1

# NLP
transformers==4.43.3
sentence-transformers==2.7.0
rapidfuzz==3.9.6
tiktoken==0.7.0

# Vector DB clients (light)
pinecone-client==4.0.0
weaviate-client==4.7.1

# LLM providers
openai==1.40.2
cohere==5.5.8

# HF hub
huggingface_hub==0.23.4

# App
streamlit==1.37.1
altair==5.3.0
